requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.35.87
botocore==1.35.87
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.5
constantly==23.10.4
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==25.1
django-storages==1.14.4
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
holidays==0.70
hyperlink==21.0.0
idna==3.10
importlib_resources==6.5.2
incremental==24.7.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.5.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.35.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.3
pillow==11.0.0
plotly==6.0.1
prompt_toolkit==3.0.51
prophet==1.1.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.4
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
stanio==0.5.1
surprise==0.1
sympy==1.13.1
textblob==0.19.0
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
